backend/requirements.txt
Django>=5.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
channels>=4.0
channels-redis>=4.2
daphne>=4.0
redis>=5.0

# MySQL (optionnel). Sous Windows, si mysqlclient échoue : pip install PyMySQL
# puis dans config/__init__.py : import pymysql; pymysql.install_as_MySQLdb()
mysqlclient>=2.2.0; platform_system != "Windows"
